+++ libs/twig/src/parser/mod.rs
@@ -1,6 +1,6 @@
+mod expression;
 pub(crate) mod general;
 mod html;
 mod twig;
-mod vue;
 
 pub(crate) use self::general::IResult;